claude 2: other context providers
