+++ src/renderer/src/components/App/AppHead.cy.tsx
@@ -275,7 +275,6 @@ describe('AppHead', () => {
         teamUpdate: mockTeamUpdate,
         resetProjectPermissions: mockResetProjectPermissions,
         lang: 'en',
-        ts: {} as any,
         resetOrbitError: cy.stub(),
         bookSuggestions: [],
         bookMap: {} as any,
@@ -300,11 +299,6 @@ describe('AppHead', () => {
         isAdmin: () => false,
         isProjectAdmin: () => false,
         flatAdd: async () => {},
-        sharedStrings: {} as any,
-        vProjectStrings: {} as any,
-        pickerStrings: {} as any,
-        projButtonStrings: {} as any,
-        newProjectStrings: {} as any,
         importOpen: false,
         setImportOpen: () => {},
         importProject: undefined,
