+++ src/renderer/src/components/App/OrgHead.cy.tsx
@@ -353,7 +353,6 @@ describe('OrgHead', () => {
         resetProjectPermissions: mockResetProjectPermissions,
         // Add other required properties to match ICtxState interface
         lang: 'en',
-        ts: {} as any,
         resetOrbitError: cy.stub(),
         bookSuggestions: [],
         bookMap: {} as any,
@@ -378,11 +377,6 @@ describe('OrgHead', () => {
         isAdmin: () => false,
         isProjectAdmin: () => false,
         flatAdd: async () => {},
-        sharedStrings: {} as any,
-        vProjectStrings: {} as any,
-        pickerStrings: {} as any,
-        projButtonStrings: {} as any,
-        newProjectStrings: {} as any,
         importOpen: false,
         setImportOpen: () => {},
         importProject: undefined,
