+++ src/renderer/src/components/Sheet/PassageCard.cy.tsx
@@ -138,7 +138,6 @@ describe('PassageCard', () => {
   ): ICtxState => ({
     t: {} as any,
     connected: true,
-    projButtonStr: {} as any,
     mediafiles: [],
     discussions: [],
     groupmemberships: [],
