+++ src/renderer/src/components/Sheet/PlanTabSelect.cy.tsx
@@ -167,7 +167,6 @@ describe('PlanTabSelect', () => {
   const createMockPlanContextState = (overrides = {}) => ({
     t: {} as any,
     connected: false,
-    projButtonStr: {} as any,
     mediafiles: [],
     discussions: [],
     groupmemberships: [],
