+++ src/renderer/src/components/Sheet/PlanView.cy.tsx
@@ -177,7 +177,6 @@ describe('PlanView', () => {
   const createMockPlanContextState = (overrides = {}) => ({
     t: {} as any,
     connected: false,
-    projButtonStr: {} as any,
     mediafiles: [],
     discussions: [],
     groupmemberships: [],
