+++ src/renderer/src/routes/ProjectsScreen.cy.tsx
@@ -183,7 +183,6 @@ describe('ProjectsScreen', () => {
     const mockTeamContextValue = {
       state: {
         lang: 'en',
-        ts: {} as any,
         resetOrbitError: cy.stub(),
         bookSuggestions: [],
         bookMap: {} as any,
@@ -211,11 +210,6 @@ describe('ProjectsScreen', () => {
         isAdmin: isAdmin,
         isProjectAdmin: () => false,
         flatAdd: async () => {},
-        sharedStrings: {} as any,
-        vProjectStrings: {} as any,
-        pickerStrings: {} as any,
-        projButtonStrings: {} as any,
-        newProjectStrings: {} as any,
         importOpen: false,
         setImportOpen: () => {},
         importProject: undefined,
